:recycle: admin: move #attachToTraining method to training adapter
